pyproject.toml
[tool.poetry]
name = "blogsai"
version = "0.1.0"
description = "AI-powered analysis of government agency press releases with enhanced research capabilities"
authors = ["Cathy <cathy.chang.ct@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
openai = "^1.3.0"
sqlalchemy = "^2.0.0"
pydantic = "^2.0.0"
pyyaml = "^6.0.0"
click = "^8.1.0"
jinja2 = "^3.1.0"
python-dateutil = "^2.8.0"
feedparser = "^6.0.0"
selenium = "^4.15.0"
schedule = "^1.2.0"
python-dotenv = "^1.0.0"
markdown = "^3.5.0"
webdriver-manager = "^4.0.0"
reportlab = "^4.0.0"
pyqt5 = "^5.15.4"
pyinstaller = "^6.0.0"
keyring = "^24.0.0"
platformdirs = "^4.0.0"
pytz = "^2023.3"


[tool.poetry.scripts]
blogsai = "blogsai.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
